requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
coverage==7.13.5
idna==3.11
iniconfig==2.3.0
lxml==4.9.4
numpy==2.4.4
packaging==26.0
pandas==2.3.3
pluggy==1.6.0
pytest==7.4.4
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.1
six==1.17.0
stadata==1.0.0
tqdm==4.67.3
tzdata==2026.1
urllib3==2.6.3
google-cloud-bigquery
flask
gunicorn
